refactor: split the functions up to different files
